+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 + Add set authorities in SecurityContext
 + Add get roles list
 + Return roles on login
++ Add check on routes for roles
 
 ## 0.14.0
 + Upgrade dependencies